feat(testing): reset database only when specified

+++ tests/TestCase.php
@@ -29,14 +29,19 @@ class TestCase extends \Illuminate\Foundation\Testing\TestCase
     {
         parent::setUp();
         $this->createApplication();
-        Artisan::call('migrate:reset');
-        Artisan::call('migrate');
-        Artisan::call('db:seed', array("--class" => "TestDatabaseSeeder"));
+
+        if (env('RESET_DATABASE_TEST', "1") === "1") {
+            Artisan::call('migrate:reset');
+            Artisan::call('migrate');
+            Artisan::call('db:seed', array("--class" => "TestDatabaseSeeder"));
+        }
     }
 
     public function tearDown()
     {
-        Artisan::call('migrate:reset');
+        if (env('RESET_DATABASE_TEST', "1") === "1") {
+            Artisan::call('migrate:reset');
+        }
         parent::tearDown();
     }
